+++ packages/core/src/executors/Parallel.ts
@@ -4,16 +4,18 @@
  */
 
 import Context from '../Context';
-import Executor from '../Executor';
+import Executor, { ExecuteHandler } from '../Executor';
 import Task from '../Task';
 
 export default class ParallelExecutor<Ctx extends Context> extends Executor<Ctx> {
+  parallel: boolean = true;
+
   /**
    * Execute tasks in parallel.
    */
-  async run<T>(tasks: Task<Ctx>[], value?: T): Promise<any[]> {
+  async run<T>(handler: ExecuteHandler<Ctx>, tasks: Task<Ctx>[], value?: T): Promise<any[]> {
     this.debug('Parallelizing %d tasks', tasks.length);
 
-    return Promise.all(tasks.map(task => this.execute(task, value, true)));
+    return Promise.all(tasks.map(task => handler(task, value)));
   }
 }