+++ packages/core/src/executors/Serial.ts
@@ -6,7 +6,7 @@
 /* eslint-disable no-restricted-syntax, no-await-in-loop */
 
 import Context from '../Context';
-import Executor from '../Executor';
+import Executor, { ExecuteHandler } from '../Executor';
 import Task from '../Task';
 
 export default class SerialExecutor<Ctx extends Context> extends Executor<Ctx> {
@@ -14,13 +14,13 @@ export default class SerialExecutor<Ctx extends Context> extends Executor<Ctx> {
    * Execute tasks in sequential order with the output of each
    * task being passed to the next promise in the chain.
    */
-  async run<T>(tasks: Task<Ctx>[], value?: T): Promise<any> {
+  async run<T>(handler: ExecuteHandler<Ctx>, tasks: Task<Ctx>[], value?: T): Promise<any> {
     this.debug('Serializing %d tasks', tasks.length);
 
     let nextValue = value;
 
     for (const task of tasks) {
-      nextValue = await this.execute(task, nextValue);
+      nextValue = await handler(task, nextValue);
     }
 
     return nextValue;