+++ packages/core/src/reporters/ErrorReporter.ts
@@ -7,6 +7,8 @@ import Reporter from '../Reporter';
 
 export default class ErrorReporter extends Reporter {
   bootstrap() {
+    super.bootstrap();
+
     this.console.on('error', this.handleError);
   }