+++ packages/core/tests/Console.test.ts
@@ -1,5 +1,5 @@
-/* eslint-disable unicorn/no-hex-escape */
-
+import chalk from 'chalk';
+import ansiEscapes from 'ansi-escapes';
 import Console from '../src/Console';
 import { createTestTool } from './helpers';
 
@@ -16,7 +16,7 @@ describe('Console', () => {
     it('writes ansi escape code', () => {
       cli.clearOutput();
 
-      expect(cli.out).toHaveBeenCalledWith('\x1Bc');
+      expect(cli.out).toHaveBeenCalledWith(ansiEscapes.eraseScreen);
     });
 
     it('resets last output height', () => {
@@ -32,7 +32,7 @@ describe('Console', () => {
       cli.lastOutputHeight = 10;
       cli.clearLinesOutput();
 
-      expect(cli.out).toHaveBeenCalledWith('\x1B[1A\x1B[K'.repeat(10));
+      expect(cli.out).toHaveBeenCalledWith(ansiEscapes.eraseLines(11));
     });
 
     it('resets last output height', () => {
@@ -337,14 +337,14 @@ describe('Console', () => {
       cli.tool.options.footer = 'Footer';
       cli.handleRender();
 
-      expect(cli.out).toHaveBeenCalledWith('');
+      expect(cli.out).toHaveBeenCalledWith(expect.not.stringContaining('Footer'));
     });
 
     it('doesnt prepend a header if not final', () => {
       cli.tool.options.header = 'Header';
       cli.handleRender();
 
-      expect(cli.out).toHaveBeenCalledWith('');
+      expect(cli.out).toHaveBeenCalledWith(expect.not.stringContaining('Header'));
     });
 
     it('doesnt display non-live logs if not final', () => {
@@ -352,7 +352,7 @@ describe('Console', () => {
       cli.errorLogs.push('Error log');
       cli.handleRender();
 
-      expect(cli.out).toHaveBeenCalledWith('');
+      expect(cli.out).toHaveBeenCalledWith(expect.not.stringContaining('Log'));
     });
 
     it('doesnt clear render listeners', () => {
@@ -417,7 +417,7 @@ describe('Console', () => {
     it('writes ansi escape code', () => {
       cli.hideCursor();
 
-      expect(cli.out).toHaveBeenCalledWith('\x1B[?25l');
+      expect(cli.out).toHaveBeenCalledWith(ansiEscapes.cursorHide);
     });
   });
 
@@ -510,15 +510,26 @@ describe('Console', () => {
     it('writes ansi escape code', () => {
       cli.resetCursor();
 
-      expect(cli.out).toHaveBeenCalledWith(expect.stringContaining('0H'));
+      expect(cli.out).toHaveBeenCalledWith(ansiEscapes.cursorTo(0, cli.size().rows));
+    });
+  });
+
+  describe('size()', () => {
+    it('returns columns and rows', () => {
+      const size = cli.size();
+
+      expect(size.columns).toBeDefined();
+      expect(size.rows).toBeDefined();
     });
   });
 
   describe('showCursor()', () => {
     it('writes ansi escape code', () => {
+      const spy = jest.spyOn(process.stdout, 'write');
+
       cli.showCursor();
 
-      expect(cli.out).toHaveBeenCalledWith('\x1B[?25h');
+      expect(spy).toHaveBeenCalledWith(ansiEscapes.cursorShow);
     });
   });
 
@@ -566,6 +577,24 @@ describe('Console', () => {
       jest.runOnlyPendingTimers();
 
       expect(renderSpy).toHaveBeenCalled();
+    });
+  });
+
+  describe('strip()', () => {
+    it('stips ANSI escape codes', () => {
+      expect(cli.strip(chalk.red('foo'))).toBe('foo');
+    });
+  });
+
+  describe('truncate()', () => {
+    it('truncates with ANSI escape codes', () => {
+      expect(cli.truncate(chalk.red('foobar'), 3)).not.toBe(chalk.red('foobar'));
+    });
+  });
+
+  describe('wrap()', () => {
+    it('wraps with ANSI escape codes', () => {
+      expect(cli.wrap(chalk.red('foobar'), 3)).toBe(chalk.red('foo\nbar'));
     });
   });